De-templatize the read-ahead driver
