+++ src/include/duckdb/parallel/scan_read_ahead.hpp
@@ -23,11 +23,11 @@
 #include "duckdb/parallel/task_executor.hpp"
 
 #include <functional>
-#include <type_traits>
 
 namespace duckdb {
 class ClientContext;
 struct TableFunctionInput;
+struct LocalTableFunctionState;
 
 class ReadAheadJobCompletion {
 public:
@@ -80,11 +80,17 @@ struct ScanReadAheadJob {
 	ScanReadAheadJob() = default;
 	ScanReadAheadJob(ScanReadAheadJob &&) noexcept = default;
 	ScanReadAheadJob &operator=(ScanReadAheadJob &&) noexcept = default;
-	~ScanReadAheadJob() {
-		// scheduled reads might still be in flight, wait for them before destroying the job
-		if (io_completion) {
-			io_completion->WaitForIO();
-		}
+	virtual ~ScanReadAheadJob();
+
+	template <class TARGET>
+	TARGET &Cast() {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<TARGET &>(*this);
+	}
+	template <class TARGET>
+	const TARGET &Cast() const {
+		DynamicCastCheck<TARGET>(this);
+		return reinterpret_cast<const TARGET &>(*this);
 	}
 
 	//! Batch index of this job, drives ordered queue admission
@@ -110,142 +116,36 @@ void ScanReadAheadYield(ClientContext &context);
 bool TryGetReadAheadDepth(ClientContext &context, optional_idx &depth);
 
 //! Drives read-ahead for a scan, its purpose is to keep several scan jobs scheduled ahead of decoding.
-//! JOB must derive from ScanReadAheadJob.
-template <class JOB, class STATE>
+//! Jobs derive from ScanReadAheadJob, callers claim them back with Cast.
 class ScanReadAhead {
 public:
 	ScanReadAhead(ClientContext &context, idx_t read_ahead_depth_p,
-	              unique_ptr<ManagedAsyncMemoryGovernor> memory_governor_p)
-	    : read_ahead_depth(read_ahead_depth_p), memory_governor(std::move(memory_governor_p)) {
-		D_ASSERT(read_ahead_depth_p > 0);
-		backlog_budget = memory_governor ? memory_governor->BackpressureBudget() : NumericLimits<idx_t>::Maximum();
-		executor = make_shared_ptr<TaskExecutor>(context, TaskSchedulerType::ASYNC);
-	}
-	~ScanReadAhead() {
-		executor->CancelAndDrain();
-	}
+	              unique_ptr<ManagedAsyncMemoryGovernor> memory_governor_p);
+	~ScanReadAhead();
 
 public:
 	//! Claims the next job and schedules its I/O, filling io_tasks when the I/O was detached to the pool.
 	//! Returns null when there are no more jobs to produce.
-	using ProduceJobCallback = std::function<unique_ptr<JOB>(vector<unique_ptr<AsyncTask>> &io_tasks)>;
+	using ProduceJobCallback = std::function<unique_ptr<ScanReadAheadJob>(vector<unique_ptr<AsyncTask>> &io_tasks)>;
 
 	//! Try to produce one job into the queue.
-	bool TryProduceJob(const ProduceJobCallback &claim_and_schedule) {
-		static_assert(std::is_base_of<ScanReadAheadJob, JOB>::value, "JOB must derive from ScanReadAheadJob");
-		ThrowIfError();
-		if (IsDone() || !TryReserveSlot()) {
-			return false;
-		}
-		ProducerReservation reservation(*this);
-		try {
-			vector<unique_ptr<AsyncTask>> io_tasks;
-			auto job = claim_and_schedule(io_tasks);
-			if (!job) {
-				// there are no more jobs to produce, the scan is done
-				SetDone();
-				return false;
-			}
-			PushJob(std::move(job), std::move(io_tasks));
-			reservation.committed = true;
-		} catch (std::exception &ex) {
-			PushError(ErrorData(ex));
-			throw;
-		} catch (...) { // LCOV_EXCL_START
-			PushError(ErrorData("Unknown exception while producing a read-ahead job"));
-			throw;
-		} // LCOV_EXCL_STOP
-		return true;
-	}
-
+	bool TryProduceJob(const ProduceJobCallback &claim_and_schedule);
 	//! Check if scan is done, i.e., no more jobs to do
-	bool IsDone() const {
-		return done.load();
-	}
-
+	bool IsDone() const;
 	//! Whether any thread holds a reserved slot it has not pushed a job for yet
-	bool HasActiveProducers() const {
-		return active_producers.load() > 0;
-	}
-
+	bool HasActiveProducers() const;
 	//! Pop the oldest queued job
-	unique_ptr<JOB> ClaimJob() {
-		lock_guard<mutex> guard(lock);
-		if (ready_queue.empty()) {
-			return nullptr;
-		}
-		auto job = std::move(ready_queue.front());
-		ready_queue.pop_front();
-		ReleaseSlot();
-		return job;
-	}
-
+	unique_ptr<ScanReadAheadJob> ClaimJob();
 	//! Push a finished job's scan state, so learned scan state carries over to jobs created later
-	void PushState(unique_ptr<STATE> state) {
-		D_ASSERT(state);
-		lock_guard<mutex> guard(lock);
-		state_pool.push_back(std::move(state));
-	}
-
+	void PushState(unique_ptr<LocalTableFunctionState> state);
 	//! Pop a recycled scan state, returns null when none is available
-	unique_ptr<STATE> TryPopState() {
-		lock_guard<mutex> guard(lock);
-		if (state_pool.empty()) {
-			return nullptr;
-		}
-		auto state = std::move(state_pool.back());
-		state_pool.pop_back();
-		return state;
-	}
-
+	unique_ptr<LocalTableFunctionState> TryPopState();
 	//! Block until the claimed job's scheduled I/O has completed
-	void WaitForJob(JOB &job) {
-		if (job.io_completion) {
-			job.io_completion->WaitForIO();
-		}
-		// the job's I/O has completed, release its budget charge
-		pending_io_bytes -= job.io_bytes;
-		job.io_bytes = 0;
-		ThrowIfError();
-	}
-
+	void WaitForJob(ScanReadAheadJob &job);
 	//! Acquire the next job, producing and claiming as needed; on ACQUIRED the job's I/O has been settled.
 	//! On PARKED the caller must yield holding the claimed job, the next call settles its completed I/O.
 	ScanReadAheadAcquire AcquireJob(ClientContext &context, TableFunctionInput &data_p,
-	                                const ProduceJobCallback &claim_and_schedule, unique_ptr<JOB> &job) {
-		if (job) {
-			// resuming with an already claimed job, settle its completed I/O
-			WaitForJob(*job);
-			job->io_completion.reset();
-			return ScanReadAheadAcquire::ACQUIRED;
-		}
-		while (true) {
-			// keep the queue full before claiming
-			while (TryProduceJob(claim_and_schedule)) {
-			}
-			job = ClaimJob();
-			if (!job && IsDone() && !HasActiveProducers()) {
-				// re-check the queue, a job may have been pushed between the claim and the done check
-				job = ClaimJob();
-				if (!job) {
-					return ScanReadAheadAcquire::EXHAUSTED;
-				}
-			}
-			if (!job) {
-				// another thread is between claiming an assignment and pushing its job, wait for it
-				ScanReadAheadYield(context);
-				continue;
-			}
-			// park until the job's scheduled I/O completes, the last I/O task to finish wakes the scan task
-			if (job->io_completion->TryPark(data_p)) {
-				return ScanReadAheadAcquire::PARKED;
-			}
-			// parking is not available to this caller or the I/O has already completed
-			WaitForJob(*job);
-			job->io_completion.reset();
-			return ScanReadAheadAcquire::ACQUIRED;
-		}
-	}
+	                                const ProduceJobCallback &claim_and_schedule, unique_ptr<ScanReadAheadJob> &job);
 
 private:
 	//! Settles the reservation taken by TryReserveSlot
@@ -264,68 +164,17 @@ private:
 	};
 
 	//! Mark the scan as done, i.e., no more jobs to produce
-	void SetDone() {
-		done = true;
-	}
-
+	void SetDone();
 	//! Reserve an in-flight job slot for producing a job
-	bool TryReserveSlot() {
-		const bool over_budget = pending_io_bytes.load() >= backlog_budget.load();
-		const idx_t depth = over_budget ? 1 : read_ahead_depth;
-		if (active_jobs.fetch_add(1) >= depth) {
-			--active_jobs;
-			return false;
-		}
-		++active_producers;
-		return true;
-	}
-
+	bool TryReserveSlot();
 	//! Schedule the job's I/O and admit the job to the queue
-	void PushJob(unique_ptr<JOB> job, vector<unique_ptr<AsyncTask>> io_tasks) {
-		// beyond its scheduled I/O a job carries scan-state overhead (row-group sized decode buffers)
-		static constexpr idx_t MINIMUM_JOB_IO_CHARGE = 16ULL * 1024 * 1024;
-		auto completion = make_shared_ptr<ReadAheadJobCompletion>(executor, io_tasks.size());
-		job->io_completion = completion;
-		for (auto &task : io_tasks) {
-			job->io_bytes += task->GetIOSize();
-		}
-		job->io_bytes = MaxValue<idx_t>(job->io_bytes, MINIMUM_JOB_IO_CHARGE);
-		pending_io_bytes += job->io_bytes;
-		// schedule the reads detached on the async pool right away
-		for (auto &task : io_tasks) {
-			executor->ScheduleTask(make_uniq<ReadAheadIOTask>(*executor, std::move(task), completion));
-		}
-		lock_guard<mutex> guard(lock);
-		if (memory_governor) {
-			memory_governor->UpdateReservation(pending_io_bytes.load());
-			backlog_budget = memory_governor->BackpressureBudget();
-		}
-		// producers push concurrently, so admit jobs to the queue in batch-index order
-		pending_jobs.emplace(job->batch_index, std::move(job));
-		while (!pending_jobs.empty() && pending_jobs.begin()->first == next_batch_index) {
-			ready_queue.push_back(std::move(pending_jobs.begin()->second));
-			pending_jobs.erase(pending_jobs.begin());
-			next_batch_index++;
-		}
-	}
-
+	void PushJob(unique_ptr<ScanReadAheadJob> job, vector<unique_ptr<AsyncTask>> io_tasks);
 	//! Push an error onto the async executor
-	void PushError(ErrorData error) {
-		executor->PushError(std::move(error));
-	}
-
+	void PushError(ErrorData error);
 	//! Throw if any read-ahead thread or task pushed an error
-	void ThrowIfError() {
-		if (executor->HasError()) {
-			executor->ThrowError();
-		}
-	}
-
+	void ThrowIfError();
 	//! Release a read-ahead slot
-	void ReleaseSlot() {
-		D_ASSERT(active_jobs.load() > 0);
-		active_jobs--;
-	}
+	void ReleaseSlot();
 
 private:
 	//! Maximum number of jobs scheduled ahead of decoding, unlimited in the -1 auto mode
@@ -336,13 +185,13 @@ private:
 	atomic<idx_t> backlog_budget {0};
 
 	mutable mutex lock;
-	deque<unique_ptr<JOB>> ready_queue;
+	deque<unique_ptr<ScanReadAheadJob>> ready_queue;
 	//! Jobs pushed out of order, held back until all earlier batch indexes are admitted to the queue
-	map<idx_t, unique_ptr<JOB>> pending_jobs;
+	map<idx_t, unique_ptr<ScanReadAheadJob>> pending_jobs;
 	//! The batch index the queue admits next
 	idx_t next_batch_index = 0;
 	//! Scan states of finished jobs
-	vector<unique_ptr<STATE>> state_pool;
+	vector<unique_ptr<LocalTableFunctionState>> state_pool;
 	//! Jobs scheduled ahead of decoding
 	atomic<idx_t> active_jobs {0};
 	//! Bytes of scheduled I/O that has not completed yet, released once the claimed job's I/O finished
